Create minecode/indexing.py #49 #285
    * Update tests
