+++ minecode/tests/test_indexing.py
@@ -10,14 +10,9 @@
 import json
 import os
 
-from mock import Mock
-from mock import patch
-
 from matchcode.models import ExactFileIndex
-from minecode.management.commands.process_scans import Command
-from minecode.management.commands.process_scans import get_scan_status
-from minecode.management.commands.process_scans import index_package_files
-from minecode.management.scanning import Scan
+
+from minecode.management import indexing
 from minecode.models import ScannableURI
 from minecode.utils_test import MiningTestCase
 from minecode.utils_test import JsonBasedTesting
@@ -25,7 +20,7 @@ from packagedb.models import Package
 from packagedb.models import Resource
 
 
-class ProcessScansTest(MiningTestCase, JsonBasedTesting):
+class IndexingTest(MiningTestCase, JsonBasedTesting):
     BASE_DIR = os.path.join(os.path.dirname(__file__), 'testfiles')
 
     def setUp(self):
@@ -37,67 +32,35 @@ class ProcessScansTest(MiningTestCase, JsonBasedTesting):
             version='20040705.181715'
         )
 
-    def test_ProcessScansTest_get_scan_status(self):
-        scan = Scan()
-        scan.status = 'not_started'
-        self.assertEqual(ScannableURI.SCAN_SUBMITTED, get_scan_status(scan))
-        scan.status = 'queued'
-        self.assertEqual(ScannableURI.SCAN_SUBMITTED, get_scan_status(scan))
-        scan.status = 'running'
-        self.assertEqual(ScannableURI.SCAN_IN_PROGRESS, get_scan_status(scan))
-        scan.status = 'success'
-        self.assertEqual(ScannableURI.SCAN_COMPLETED, get_scan_status(scan))
-        scan.status = 'failure'
-        self.assertEqual(ScannableURI.SCAN_FAILED, get_scan_status(scan))
-        scan.status = 'stopped'
-        self.assertEqual(ScannableURI.SCAN_FAILED, get_scan_status(scan))
-        scan.status = 'stale'
-        self.assertEqual(ScannableURI.SCAN_FAILED, get_scan_status(scan))
-        scan.status = 'asdf'
-        self.assertRaises(Exception, get_scan_status, scan)
-
     def test_ProcessScansTest_index_package_files(self):
         scan_data_loc = self.get_test_loc('scancodeio/get_scan_data.json')
         with open(scan_data_loc, 'rb') as f:
             scan_data = json.loads(f.read())
-        self.assertEqual(0, len(index_package_files(self.package1, scan_data)))
+        self.assertEqual(0, len(indexing.index_package_files(self.package1, scan_data)))
         result = Resource.objects.filter(package=self.package1)
         self.assertEqual(64, len(result))
         results = [r.to_dict() for r in result]
         expected_resources_loc = self.get_test_loc('scancodeio/get_scan_data_expected_resources.json')
         self.check_expected_results(results, expected_resources_loc, regen=False)
 
-    @patch('requests.get')
-    def test_ProcessScansTest_process_scan(self, mock_get):
-        # Set up mock responses
-        mock_scan_info_response = Mock()
-        scan_info_loc = self.get_test_loc('scancodeio/get_scan_info.json')
-        with open(scan_info_loc, 'rb') as f:
-            mock_scan_info_response.json.return_value = json.loads(f.read())
-
-        mock_scan_data_response = Mock()
+    def test_ProcessScansTest_process_scan(self):
         scan_data_loc = self.get_test_loc('scancodeio/get_scan_data.json')
         with open(scan_data_loc, 'rb') as f:
-            mock_scan_data_response.json.return_value = json.loads(f.read())
+            scan_data = json.load(f)
 
-        mock_scan_summary_response = Mock()
         scan_summary_loc = self.get_test_loc('scancodeio/scan_summary_response.json')
         with open(scan_summary_loc, 'rb') as f:
-            mock_scan_summary_response.json.return_value = json.loads(f.read())
-
-        mock_get.side_effect = [mock_scan_info_response, mock_scan_data_response, mock_scan_summary_response]
+            scan_summary = json.load(f)
 
         # Set up ScannableURI
-        scan_uuid = '54dc4afe-70ea-4f1c-9ed3-989efd9a991f'
         scannable_uri = ScannableURI.objects.create(
             uri='https://repo1.maven.org/maven2/maven/wagon-api/20040705.181715/wagon-api-20040705.181715.jar',
-            scan_uuid=scan_uuid,
             scan_status=ScannableURI.SCAN_COMPLETED,
             package=self.package1
         )
 
         # Run test
-        Command.process_scan(scannable_uri)
+        indexing.index_package(scannable_uri, self.package1, scan_data, scan_summary)
 
         # Make sure that we get license_expression and copyright from the summary
         self.assertEqual('apache-2.0', self.package1.declared_license_expression)
@@ -107,4 +70,3 @@ class ProcessScansTest(MiningTestCase, JsonBasedTesting):
         self.assertEqual(64, result.count())
         result = ExactFileIndex.objects.filter(package=self.package1)
         self.assertEqual(45, result.count())
-
